+++ Source/CoachMarkBodyDefaultView.swift
@@ -116,7 +116,7 @@ public class CoachMarkBodyDefaultView : UIControl, CoachMarkBodyView {
         hintLabel.font = UIFont.systemFontOfSize(15.0)
         hintLabel.scrollEnabled = false
         hintLabel.textAlignment = .Justified
-        hintLabel.layoutManager.hyphenationFactor = 2.0
+        hintLabel.layoutManager.hyphenationFactor = 1.0
         hintLabel.editable = false
 
         nextLabel.textColor = UIColor.darkGrayColor()
@@ -169,7 +169,7 @@ public class CoachMarkBodyDefaultView : UIControl, CoachMarkBodyView {
         hintLabel.font = UIFont.systemFontOfSize(15.0)
         hintLabel.scrollEnabled = false
         hintLabel.textAlignment = .Justified
-        hintLabel.layoutManager.hyphenationFactor = 2.0
+        hintLabel.layoutManager.hyphenationFactor = 1.0
         hintLabel.editable = false
         
         hintLabel.translatesAutoresizingMaskIntoConstraints = false