+++ Source/CoachMarkDisplayManager.swift
@@ -1,6 +1,6 @@
 // CoachMarkDisplayManager.swift
 //
-// Copyright (c) 2015 Frédéric Maquin <fred@ephread.com>
+// Copyright (c) 2015, 2016 Frédéric Maquin <fred@ephread.com>
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -20,7 +20,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-import Foundation
+import UIKit
 
 /// This class deals with the layout of coach marks.
 internal class CoachMarkDisplayManager {
@@ -90,6 +90,7 @@ internal class CoachMarkDisplayManager {
 
         self.prepareCoachMarkForDisplay()
         self.overlayView.disableOverlayTap = coachMark.disableOverlayTap
+        self.overlayView.allowTouchInsideCutoutPath = coachMark.allowTouchInsideCutoutPath
 
         self.clearStoredData()