+++ Source/Instructions.swift
@@ -1,6 +1,6 @@
 // Instructions.swift
 //
-// Copyright (c) 2015 Frédéric Maquin <fred@ephread.com>
+// Copyright (c) 2015, 2016 Frédéric Maquin <fred@ephread.com>
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -20,8 +20,10 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-import Foundation
+import UIKit
 
-let kOverlayFadeAnimationDuration: NSTimeInterval = 0.3
-let kCoachMarkFadeAnimationDuration: NSTimeInterval = 0.3
-let kOverlayColor = UIColor(red: 226.0/255.0, green: 226.0/255.0, blue: 226.0/255.0, alpha: 0.65)
+struct Constants {
+    static let overlayFadeAnimationDuration: NSTimeInterval = 0.3
+    static let coachMarkFadeAnimationDuration: NSTimeInterval = 0.3
+    static let overlayColor = UIColor(red: 226.0/255.0, green: 226.0/255.0, blue: 226.0/255.0, alpha: 0.65)
+}