+++ Source/OverlayView.swift
@@ -28,7 +28,7 @@ internal class OverlayView: UIView {
     //MARK: - Internal properties
 
     /// The background color of the overlay
-    var overlayColor: UIColor = kOverlayColor
+    var overlayColor: UIColor = Constants.overlayColor
 
     /// The blur effect style to apply to the overlay.
     /// Setting this property to anything but `nil` will
@@ -63,6 +63,9 @@ internal class OverlayView: UIView {
     /// Used to temporarily disable the tap, for a given coachmark.
     var disableOverlayTap: Bool = false
 
+    /// Used to temporarily enable touch forwarding isnide the cutoutPath.
+    var allowTouchInsideCutoutPath: Bool = false
+
     /// Delegate to which tell that the overlay view received a tap event.
     weak var delegate: OverlayViewDelegate?
 
@@ -85,7 +88,7 @@ internal class OverlayView: UIView {
 
     /// TapGestureRecognizer that will catch tap event performed on the overlay
     private lazy var singleTapGestureRecognizer: UITapGestureRecognizer = {
-        let gestureRecognizer = UITapGestureRecognizer(target: self, action: "handleSingleTap:")
+        let gestureRecognizer = UITapGestureRecognizer(target: self, action: #selector(OverlayView.handleSingleTap(_:)))
 
         return gestureRecognizer
     }()
@@ -216,6 +219,10 @@ internal class OverlayView: UIView {
 
         if hitView == self {
             guard let cutoutPath = self.cutoutPath else {
+                return hitView
+            }
+
+            if !self.allowTouchInsideCutoutPath {
                 return hitView
             }